requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=2.1.0
python-dotenv>=1.0.0
